Use explicit casts

+++ GHMarkdownParser/GHMarkdownParser/GHMarkdownParser.m
@@ -15,7 +15,7 @@
     NSString *HTMLString = nil;
     
     char *markdownUTF8 = (char*)markdownString.UTF8String;
-    Document *document = mkd_string(markdownUTF8, strlen(markdownUTF8), 0);
+    Document *document = mkd_string(markdownUTF8, (int)strlen(markdownUTF8), 0);
     if (document && mkd_compile(document, 0)) {
         char *HTMLUTF8 = NULL;
         int length = mkd_document(document, &HTMLUTF8);
